refactor for schematic editor
