+++ model_railway_signals/library/common.py
