+++ model_railway_signals/library/resources/bell-ring-01.wav
