+++ model_railway_signals/library/resources/bell-ring-02.wav
