+++ model_railway_signals/library/resources/bell-ring-03.wav
