+++ model_railway_signals/library/resources/telegraph-key-01.wav
