+++ model_railway_signals/library/signals.py
