+++ model_railway_signals/library/signals_colour_lights.py
