+++ model_railway_signals/library/signals_ground_disc.py
