+++ model_railway_signals/library/signals_semaphores.py
