+++ model_railway_signals/library/track_sections.py
