+++ model_railway_signals/library/track_sensors.py
